changed constants location
