+++ src/js/components/overview/charts/BentoBarChart.js
@@ -2,8 +2,13 @@ import React from 'react';
 
 import { BarChart, Bar, Cell, XAxis, YAxis, Tooltip, Label } from 'recharts';
 
-import { BAR_CHART_FILL, BAR_CHART_MISSING_FILL } from '../../../constants/overviewConstants';
-import { TOOL_TIP_STYLE, COUNT_STYLE, LABEL_STYLE } from './common/styles';
+import {
+  BAR_CHART_FILL,
+  BAR_CHART_MISSING_FILL,
+  TOOL_TIP_STYLE,
+  COUNT_STYLE,
+  LABEL_STYLE,
+} from '../../../constants/overviewConstants';
 
 const ASPECT_RATIO = 1.2;
 const MAX_TICK_LABEL_CHARS = 15;
