--- src/js/components/overview/charts/common/styles.js
@@ -1,22 +1,0 @@
-export const TOOL_TIP_STYLE = {
-  backgroundColor: 'rgba(255, 255, 255, 0.9)',
-  padding: '5px',
-  border: '1px solid grey',
-  boxShadow: '0px 0px 2px rgba(0, 0, 0, 0.9)',
-  borderRadius: '2px',
-  textAlign: 'left',
-};
-
-export const LABEL_STYLE = {
-  fontWeight: 'bold',
-  fontSize: '12px',
-  padding: '0',
-  margin: '0',
-};
-
-export const COUNT_STYLE = {
-  fontWeight: 'normal',
-  fontSize: '11px',
-  padding: '0',
-  margin: '0',
-};
